requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
dostoevsky==0.6.0
fasttext==0.9.2
numpy==1.21.2
pybind11==2.8.0
python-telegram-bot==13.7
pytz==2021.3
razdel==0.5.0
six==1.16.0
tornado==6.1
tzdata==2021.2.post0
tzlocal==3.0
